+++ packages/yew/src/app_handle.rs
@@ -25,7 +25,11 @@ where
     pub(crate) fn mount_with_props(element: Element, props: Rc<COMP::Properties>) -> Self {
         clear_element(&element);
         let app = Self {
-            scope: Scope::new(None),
+            scope: Scope::new(
+                None,
+                #[cfg(debug_assertions)]
+                u64::MAX,
+            ),
         };
         app.scope
             .mount_in_place(element, NodeRef::default(), NodeRef::default(), props);
